#19 1.created seperate test files for each API
